backend/requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.32.1
pydantic==2.10.4
sqlalchemy==2.0.34
alembic==1.16.2
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.28.1
chromadb==1.0.13
pandas==2.2.3
openpyxl==3.1.2
numpy>=1.26.0
scikit-learn==1.5.2
python-socketio==5.11.1
redis==5.2.1
celery==5.4.0
pytest==8.3.4
pytest-asyncio==0.24.0 